Finished shared_data.hpp: without testing

+++ hdrs/queue.hpp
@@ -6,19 +6,19 @@
 /*   By: blackrider <blackrider@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 07:56:22 by BlackRider        #+#    #+#             */
-/*   Updated: 2025/07/09 08:35:23 by blackrider       ###   ########.fr       */
+/*   Updated: 2025/07/10 14:38:35 by blackrider       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cstdint>
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 class FSQueue
 {
 	private:
-		uint16_t	head;
-		uint16_t	tail;
-		uint16_t	count;
+		uint8_t	head;
+		uint8_t	tail;
+		uint8_t	count;
 		data_t 		data[size];
 	public:
 		FSQueue() : head(0), tail(0), count(0) {}
@@ -27,12 +27,12 @@ class FSQueue
 		bool	pop(data_t &item);
 		bool	peek(data_t &item) const;
 		bool	peek_second(data_t &item) const;
-		uint16_t get_count() const;
+		uint8_t get_count() const;
 		bool is_empty() const;
 		bool is_full() const;
 };
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::push(const data_t &item)
 {
 	bool result = false;
@@ -47,7 +47,7 @@ bool FSQueue<data_t, size>::push(const data_t &item)
 	return result;
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::pop(data_t &item)
 {
 	bool result = false;
@@ -62,7 +62,7 @@ bool FSQueue<data_t, size>::pop(data_t &item)
 	return result;
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::pop()
 {
 	bool result = false;
@@ -76,7 +76,7 @@ bool FSQueue<data_t, size>::pop()
 	return result;
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::peek(data_t& item) const
 {
 	bool	result = false;
@@ -89,7 +89,7 @@ bool FSQueue<data_t, size>::peek(data_t& item) const
 	return result;
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::peek_second(data_t& item) const
 {
 	bool	result = false;
@@ -102,19 +102,19 @@ bool FSQueue<data_t, size>::peek_second(data_t& item) const
 	return result;
 }
 
-template <typename data_t, uint16_t size>
-uint16_t FSQueue<data_t, size>::get_count() const
+template <typename data_t, uint8_t size>
+uint8_t FSQueue<data_t, size>::get_count() const
 {
 	return count;
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::is_empty() const
 {
 	return (count == 0);
 }
 
-template <typename data_t, uint16_t size>
+template <typename data_t, uint8_t size>
 bool FSQueue<data_t, size>::is_full() const
 {
 	return (count == size);
